Fix comparison warning on STM32L4 devices
Cast the pointer used in l4_retarget to uint32_t before comparing it
to fix the warning:
"comparison between pointer and integer"

+++ targets/TARGET_STM/TARGET_STM32L4/l4_retarget.c
@@ -51,7 +51,7 @@ void *__wrap__sbrk(int incr)
     uint32_t heap_ind_old = heap_ind;
     uint32_t heap_ind_new = heap_ind_old + incr;
 
-    if (heap_ind_new > &__mbed_krbs_start) {
+    if (heap_ind_new > (uint32_t)&__mbed_krbs_start) {
         errno = ENOMEM;
         return (void *) - 1;
     }
